feat!: correct update year command for license files
`update_year`'s old behavior was to simply overwrite the current year on
file which is incorrect. This command now correctly updates the copyright so
retaining first copyright year set and appending the succeeding years
correctly.

For example, if the year on the copyright file is `2020` and the current
year is `2021`, then the correct modification should be `2020-2021`.
Here are more conversions. For all examples, the current year is `2021`.

`2020` -> `2020-2021`
`2019` -> `2019, 2021`
`2017, 2019` -> `2017, 2019, 2021`
`2018, 2020` -> `2018, 2020-2021`

Fixes #8.